squash! entangle: Fix build with glib 2.82
`glib-2.0.pc` now has `libsysprof-capture-4` in `Requires.private` field.
Normally, this does not matter since we patch our `pkg-config` to ignore that
outside of static builds but `entangle` was using `pkgconf`, which is not patched.

Let’s switch to standard `pkg-config`.
This will also allow us to clean some more dependencies.

Unfortunately, this also reveals issue with an underspecified `gio-unix-2.0` dependency so we need to fix that.